Splits webpack config
