+++ server/views/template.js
@@ -5,11 +5,11 @@ export default ({ body, title }) => {
       <head>
         <title>${title}</title>
         <link rel="stylesheet" href="/public/bundle.css">
-        <script src="/public/bundle.js"></script>
-      </head>
-
-      <body>
+        </head>
+        
+        <body>
         <div id="root">${body}</div>
+        <script src="/public/bundle.js"></script>
       </body>
     </html>
   `;
